Database now being initialized for storing users and api's
